requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
lxml==5.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyYAML==6.0.1
requests==2.31.0
urllib3==2.1.0
